\name{IceFloe}
\docType{data}
\alias{IceFloe}
\alias{icefloe}
\title{Image of ice floes}
\description{
This data represents a binary matrix, representing an image of ice floes.
}
\usage{IceFloe}
\format{A matrix containing 40 rows and 40 columns}
\source{Banfield, J. and Raftery, A. (1992). Ice floe identification in satellite images using mathematical
morphology and clustering about principal curves. Journal of the American Statistical Association,
87(417):7-16.}

\keyword{datasets}


