\docType{data}
\name{Pollution}
\alias{pollution}
\alias{Pollution}
\title{Pollution Data}
\format{A matrix containing 60 rows and 16 columns}
\source{
  McDonald, G.C. and Schwing, R.C. (1973) 'Instabilities of
  regression estimates relating air pollution to
  mortality', Technometrics, vol.15, 463-482.
}
\description{
  This data contains 1 response (mortality, normalized to
  have mean zero) along with 15 pollution-related
  explanatory variables.
}
\keyword{datasets}

