\docType{data}
\name{IceFloe}
\alias{icefloe}
\alias{IceFloe}
\title{Image of ice floes}
\format{A matrix containing 40 rows and 40 columns}
\source{
  Banfield, J. and Raftery, A. (1992). Ice floe
  identification in satellite images using mathematical
  morphology and clustering about principal curves. Journal
  of the American Statistical Association, 87(417):7-16.
}
\description{
  This data represents a binary matrix, representing an
  image of ice floes.
}
\keyword{datasets}

